Remove all the redundant presenters. #52
