+++ app-domain/src/main/java/com/paper/domain/useCase/LoadPaperAndBindModel.kt
@@ -23,7 +23,7 @@
 package com.paper.domain.useCase
 
 import com.paper.domain.event.ProgressEvent
-import com.paper.domain.widget.canvas.IPaperWidget
+import com.paper.domain.widget.editor.IPaperWidget
 import com.paper.model.repository.IPaperRepo
 import io.reactivex.*
 import io.reactivex.disposables.Disposable
