--- app-domain/src/main/java/com/paper/domain/widget/canvas/IScrapWidget.kt
@@ -1,42 +1,0 @@
-// Copyright Feb 2018-present boyw165@gmail.com
-//
-// Permission is hereby granted, free of charge, to any person obtaining
-// a copy of this software and associated documentation files (the "Software"),
-// to deal in the Software without restriction, including without limitation
-// the rights to use, copy, modify, merge, publish, distribute, sublicense,
-// and/or sell copies of the Software, and to permit persons to whom the
-// Software is furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included
-// in all copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
-// OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
-// THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
-// FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
-// DEALINGS IN THE SOFTWARE.
-
-package com.paper.domain.widget.canvas
-
-import com.paper.domain.event.DrawSVGEvent
-import com.paper.model.ScrapModel
-import com.paper.model.TransformModel
-import io.reactivex.Observable
-import java.util.*
-
-interface IScrapWidget: IBaseWidget<ScrapModel> {
-
-    // For input //////////////////////////////////////////////////////////////
-
-    fun handleTap(x: Float, y: Float)
-
-    // For output /////////////////////////////////////////////////////////////
-
-    fun getId(): UUID
-
-    fun onDrawSVG(): Observable<DrawSVGEvent>
-
-    fun onTransform(): Observable<TransformModel>
-}
