+++ app-domain/src/test/java/com/paper/domain/LoadPaperAndBindModelTest.kt
@@ -22,7 +22,7 @@ package com.paper.domain
 
 import com.paper.domain.event.ProgressEvent
 import com.paper.domain.useCase.LoadPaperAndBindModel
-import com.paper.domain.widget.canvas.IPaperWidget
+import com.paper.domain.widget.editor.IPaperWidget
 import com.paper.model.PaperModel
 import com.paper.model.repository.IPaperRepo
 import io.reactivex.Single
