+++ app-presentation/src/main/res/layout/activity_paper_editor.xml
@@ -19,7 +19,7 @@
         android:layout_width="0dp"
         android:layout_height="0dp"
         android:paddingTop="@dimen/toolbar_height"
-        app:layout_constraintBottom_toTopOf="@+id/editing_panel"
+        app:layout_constraintBottom_toTopOf="@+id/edit_panel"
         app:layout_constraintHeight_default="spread"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
@@ -80,7 +80,7 @@
 
     <!-- Editor panel -->
     <com.paper.view.editPanel.PaperEditPanelView
-        android:id="@+id/editing_panel"
+        android:id="@+id/edit_panel"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         app:layout_constraintBottom_toBottomOf="parent"
